Add initial firmware (#1)
This initial firmware serves as a sanity check. It is able to detect when the magnetic reed switch is triggered and triggers the built in LED.

Future work will track unique revolutions, which can be used to populate information on a OLED display (distance, duration, etc) and also be written as a CSV to a microSD card. Since an ESP8266 WiFi microcontroller is used, we could also dump data to a simple server.