Local server, web client, firmware to get telemetry visible on an iPad (#6)
This change adds an extremely barebones basic server and web client I run on my local network, that the ESP8266 sends revolution samples to. Ultimately, this lays the groundwork for me to view my bike telemetry on my iPad while I bike. I hacked most of this out quickly on Saturday evening, so figured I'd get it merged and improve on it as I go.

The ESP8266 firmware has been modified to make requests to this server and send batches of revolution data with relative timestamps in milliseconds. 

The web server doesn't use a database for now--just reads and writes to files on disk. This is fast enough and simple enough for my at-home use case, and for now I'm just going for whatever is easiest. 

This works end to end now--riding the bike causes revolution data to be recorded over time, which shows up in the web client (viewed on an iPad). 


![bike_gif2](https://user-images.githubusercontent.com/6299853/106572591-781b1680-64ed-11eb-8925-90110f0170e8.gif)